backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
Django==3.2.18
djangorestframework==3.13.1
djongo==1.3.6
dnspython==2.4.2
pymongo==3.12.3
pytz==2025.2
sqlparse==0.2.4
tzdata==2025.2
django-cors-headers==3.13.0